+++ drivers/display/CMakeLists.txt
@@ -4,7 +4,7 @@ zephyr_library()
 zephyr_library_sources_ifdef(CONFIG_DISPLAY_MCUX_ELCDIF	display_mcux_elcdif.c)
 zephyr_library_sources_ifdef(CONFIG_DISPLAY_NRF_LED_MATRIX display_nrf_led_matrix.c)
 zephyr_library_sources_ifdef(CONFIG_DUMMY_DISPLAY	display_dummy.c)
-zephyr_library_sources_ifdef(CONFIG_FRAMEBUF_DISPLAY	display_framebuf.c)
+zephyr_library_sources_ifdef(CONFIG_INTEL_MULTIBOOTFB_DISPLAY	display_intel_multibootfb.c)
 zephyr_library_sources_ifdef(CONFIG_UC81XX		uc81xx.c)
 zephyr_library_sources_ifdef(CONFIG_ILI9XXX		display_ili9xxx.c)
 zephyr_library_sources_ifdef(CONFIG_ILI9340		display_ili9340.c)
