+++ drivers/display/Kconfig
@@ -35,11 +35,6 @@ source "drivers/display/Kconfig.dummy"
 source "drivers/display/Kconfig.ls0xx"
 source "drivers/display/Kconfig.rm68200"
 source "drivers/display/Kconfig.max7219"
-
-config FRAMEBUF_DISPLAY
-	# Hidden, selected by client drivers.
-	bool
-	help
-	  Enable framebuffer-based display 'helper' driver.
+source "drivers/display/Kconfig.intel_multibootfb"
 
 endif # DISPLAY
