+++ include/zephyr/arch/x86/multiboot.h
@@ -96,7 +96,7 @@ struct multiboot_mmap {
 #define MULTIBOOT_HEADER_FLAG_MEM	BIT(1)	/* want mem_/mmap_* info */
 #define MULTIBOOT_HEADER_FLAG_FB	BIT(2)	/* want fb_* info */
 
-#ifdef CONFIG_MULTIBOOT_FRAMEBUF
+#ifdef CONFIG_INTEL_MULTIBOOTFB_DISPLAY
 #define MULTIBOOT_HEADER_FLAGS \
 	(MULTIBOOT_HEADER_FLAG_FB | MULTIBOOT_HEADER_FLAG_MEM)
 #else
