+++ tests/arch/x86/info/prj.conf
@@ -1,7 +1,7 @@
 CONFIG_ACPI=y
 CONFIG_MULTIBOOT_INFO=y
 CONFIG_MULTIBOOT_MEMMAP=y
-CONFIG_MULTIBOOT_FRAMEBUF=y
 CONFIG_COUNTER=y
 CONFIG_COUNTER_CMOS=y
+CONFIG_DISPLAY=y
 CONFIG_X86_MEMMAP=y
